requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-storage
google-cloud-aiplatform
